third_party/blink/web_tests/css3/calc/cross-fade-calc.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: calc() function in webkit-cross-fade</title>
<link rel="author" title="Chris Nardi" href="mailto:cnardi@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-values/#funcdef-calc">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
    #test {
        background-image: -webkit-cross-fade(url(dummy://example.png), url(dummy://example.png), calc(35% * 2));
    }
</style>
<div id="test"></div>
<script>
    const div = document.querySelector("#test");
    const expected = '-webkit-cross-fade(url("dummy://example.png"), url("dummy://example.png"), 70%)'
    test(function() {
        assert_equals(getComputedStyle(div).backgroundImage, expected);
    }, "calc() in webkit-cross-fade");
</script>
